add checking file before create link

+++ LinkProjectToSamples.py
@@ -77,8 +77,10 @@ def updateRun(runs, reqID, sample):
             for possibleRun in runID:
                 source_path = DELIVERY + "FASTQ/{}/Project_{}/{}".format(source, reqID, sample)
                 possibleRun_path = DELIVERY + "FASTQ/{}/Project_{}/{}".format(possibleRun, reqID, sample)
-                if os.path.getmtime(possibleRun_path) > os.path.getmtime(source_path):
-                    source = possibleRun
+                # check if folder exists before create link for cases that project contains old samples eg: 08822
+                if os.path.exists(source_path) and os.path.exists(possibleRun_path):
+                    if os.path.getmtime(possibleRun_path) > os.path.getmtime(source_path):
+                        source = possibleRun
             updatedRuns.append(source)
 
     return updatedRuns
@@ -162,10 +164,13 @@ def link_by_request(reqID):
                     madeDir.append(dlink)
                     subprocess.run(cmd, shell=True)
                 slink = FASTQ_ROOT % (run, reqID, sample)
-
-                cmd = "ln -sf {} {}".format(slink, dlink)
-                print (cmd)
-                subprocess.run(cmd, shell=True)
+                # check if folder exists before create link for cases that project contains old samples eg: 08822
+                if os.path.exists(slink):
+                    cmd = "ln -sf {} {}".format(slink, dlink)
+                    print (cmd)
+                    subprocess.run(cmd, shell=True)
+                else:
+                    print("{} not exits".format(slink))
 
     setaccess.set_request_acls(reqID, "")
